Reorganize the Modules archieve and rename the userUseCase to UserFormUseCase
