+++ app/src/test/java/app/igormatos/botaprarodar/data/repository/UserFormRepositoryTest.kt
@@ -3,6 +3,7 @@ package app.igormatos.botaprarodar.data.repository
 import app.igormatos.botaprarodar.data.network.api.UserApi
 import app.igormatos.botaprarodar.domain.model.AddDataResponse
 import app.igormatos.botaprarodar.utils.userRequest
+import com.google.firebase.database.FirebaseDatabase
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
@@ -13,11 +14,12 @@ import org.junit.Test
 class UserFormRepositoryTest {
 
     private var mockApi = mockk<UserApi>()
+    private var firebaseDatabase = mockk<FirebaseDatabase>()
     private lateinit var userRepository: UserRepository
 
     @Before
     fun setup() {
-        userRepository = UserRepository(userApi = mockApi)
+        userRepository = UserRepository(userApi = mockApi, firebaseDatabase = firebaseDatabase)
     }
 
     @Test
