+++ app/src/test/java/app/igormatos/botaprarodar/presentation/userForm/UserFormViewModelTest.kt
@@ -1,12 +1,11 @@
-package app.igormatos.botaprarodar.presentation.user_form
+package app.igormatos.botaprarodar.presentation.userForm
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import app.igormatos.botaprarodar.R
 import app.igormatos.botaprarodar.common.ViewModelStatus
 import app.igormatos.botaprarodar.domain.model.User
 import app.igormatos.botaprarodar.domain.model.community.Community
-import app.igormatos.botaprarodar.domain.usecase.user.UserUseCase
-import app.igormatos.botaprarodar.presentation.userForm.UserFormViewModel
+import app.igormatos.botaprarodar.domain.usecase.userForm.UserFormUseCase
 import app.igormatos.botaprarodar.utils.userFake
 import com.brunotmgomes.ui.SimpleResult
 import io.mockk.coEvery
@@ -22,7 +21,7 @@ class UserFormViewModelTest {
     @get:Rule
     var instantExecutorRule = InstantTaskExecutorRule()
 
-    private val userUseCase = mockk<UserUseCase>()
+    private val userUseCase = mockk<UserFormUseCase>()
     private val community = mockk<Community>(relaxed = true)
     private lateinit var formViewModel: UserFormViewModel
